@schnerd 2020-11-03
@MarcelGerber Huge find, thank you for digging into this. Actually surprised that it worked before, but I guess this was a bug that only impacted performance but not correctness?