@MarcelGerber 2020-11-03
Yep I think so! It was only impacting performance by doing way too many runs, but it still converged to the same end result.